Split Walker class
