+++ plextraktsync/sync.py
@@ -13,9 +13,9 @@ if TYPE_CHECKING:
 
     from plextraktsync.config.SyncConfig import SyncConfig
     from plextraktsync.media import Media
+    from plextraktsync.plan.Walker import Walker
     from plextraktsync.plex.PlexApi import PlexApi
     from plextraktsync.trakt.TraktApi import TraktApi
-    from plextraktsync.walker import Walker
 
 
 class Sync:
